Made the script into a load script instead of run script
Instead of a [Start] .NET Attribute, we now use two: The [DeclareMenuAction] which allows us to define a new Menu point within the EPLAN's menu structure. In our case, very easy, it will show under the Utilities menu.

The other Attribute is the one that allows us to define the actual registered name of our Action. Here, I chose to name it the "ToggleUnitsAction".

Please note: The name of the Attribute does not have to end with "Action". This is more a personal coding style than anything.